Allow six to be unbundled.
Since six is a bit complicated, this is based on astropy/astropy#2623.
